Refactored ChessData\Cli\Seed\Openings
